FSI: shift linear algebra functionality to a separate file
